# expression: implement vectorized evaluation for `builtinArithmeticModIntSig`